backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
pydantic==2.5.3
supabase==2.3.4
python-dotenv==1.0.1
